refactor: lighting dir
